% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnet.R
\name{resnet_v1}
\alias{resnet_v1}
\title{ResNet Version 1 Model}
\usage{
resnet_v1(input_shape, depth, num_filters = 16, num_classes = 10)
}
\arguments{
\item{input_shape}{shape of input image tensor}

\item{depth}{number of core convolutional layers}

\item{num_filters}{initial number of filters}

\item{num_classes}{number of classes}
}
\description{
Stacks of 2 x (3 x 3) Conv2D-BN-ReLU
Last ReLU is after the shortcut connection.
At the beginning of each stage, the feature map size is halved (downsampled)
by a convolutional layer with strides=2, while the number of filters is
doubled. Within each stage, the layers have the same number filters and the
same number of filters.
Features maps sizes:
stage 0: 32x32, 16
stage 1: 16x16, 32
stage 2:  8x8,  64
The Number of parameters is approx the same as Table 6 of the ResNet paper:
ResNet20 0.27M
ResNet32 0.46M
ResNet44 0.66M
ResNet56 0.85M
ResNet110 1.7M
}
\references{
Kaiming He, Xiangyu Zhang, Shaoqing Ren, and Jian
Sun. Deep residual learning for image recognition.
}
