% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnet.R
\name{resnet_v2}
\alias{resnet_v2}
\title{ResNet Version 2 Model}
\usage{
resnet_v2(input_shape, depth, num_filters = 16, num_classes = 10)
}
\arguments{
\item{input_shape}{shape of input image tensor}

\item{depth}{number of core convolutional layers}

\item{num_filters}{initial number of filters}

\item{num_classes}{number of classes}
}
\description{
Stacks of (1 x 1)-(3 x 3)-(1 x 1) BN-ReLU-Conv2D or also known as
bottleneck layer
First shortcut connection per layer is 1 x 1 Conv2D.
Second and onwards shortcut connection is identity.
At the beginning of each stage, the feature map size is halved (downsampled)
by a convolutional layer with strides=2, while the number of filter maps is
doubled. Within each stage, the layers have the same number filters and the
same filter map sizes.
Features maps sizes:
  conv1  : 32x32,  16
stage 0: 32x32,  64
stage 1: 16x16, 128
stage 2:  8x8,  256
}
\references{
He, Kaiming & Zhang, Xiangyu & Ren, Shaoqing & Sun, Jian. (2016). Identity Mappings
in Deep Residual Networks. 9908. 630-645. 10.1007/978-3-319-46493-0_38.
}
